refactor(demo): `index` is the way in, `Button` gets its own route
The home screen was the Button verification screen with a link to
`PressableFeedback` bolted on top of it, so every new component meant either a
longer scroll or a second convention. It is now a list of routes and nothing
else: adding a component is a route plus one line in `SCREENS`.

The Button screen moves to `/button` unchanged — the sections, the helpers and
the press counter are the same file, minus the link that now belongs to the
hub. `_layout` names the new route, and `index` takes the app's own title.
