+++ apps/demo/app/_layout.tsx
@@ -21,7 +21,8 @@ export default function RootLayout() {
       <XAUIProvider colorMode={colorScheme === 'dark' ? 'dark' : 'light'}>
         <PortalHost>
           <Stack>
-            <Stack.Screen name="index" options={{ title: 'Button (v1)' }} />
+            <Stack.Screen name="index" options={{ title: 'XAUI (v1)' }} />
+            <Stack.Screen name="button" options={{ title: 'Button (v1)' }} />
             <Stack.Screen
               name="pressable-feedback"
               options={{ title: 'PressableFeedback (v1)' }}
